+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/Lab.kt
@@ -2,7 +2,7 @@ package dev.kdrag0n.android12ext.monet.colors
 
 // Interface for Lab complementary color spaces
 interface Lab : Color {
-    val L: Float
-    val a: Float
-    val b: Float
+    val L: Double
+    val a: Double
+    val b: Double
 }
